refactor appType to platform
