+++ src/client/src/store/config/combineReducers.ts
@@ -4,7 +4,7 @@ import previewStatus from "./config/previewReducer";
 import validations from "./validations/setValidationsReducer";
 import versions from "./versions/reducer";
 import { azureProfileData } from "./azure/reducer";
-import appType from "./app/reducer";
+import platform from "./platform/reducer";
 
 const configReducer = combineReducers({
   detailsPage,
@@ -12,7 +12,7 @@ const configReducer = combineReducers({
   validations,
   versions,
   azureProfileData,
-  appType
+  platform
 });
 
 export default configReducer;
