+++ src/client/src/store/config/configType.ts
@@ -1,4 +1,4 @@
-import {ISelectProjectTypeAction} from "./app/model";
+import {ISelectProjectTypeAction} from "./platform/model";
 import { IVersionData } from "./versions/model";
 import { IResetWizardAction, IPreviewStatusActionTypeAction } from "./config/model";
 import { IProjectPathValidationAction, IValidationsAction} from "./validations/model";
