+++ src/client/src/store/config/platform/actionTypes.ts
