+++ src/client/src/store/config/platform/model.ts
@@ -1,7 +1,6 @@
-import { ISelected } from "../../../types/selected";
 import { CONFIG_TYPEKEYS } from "../configTypeKeys";
 
 export interface ISelectProjectTypeAction {
   type: CONFIG_TYPEKEYS.SELECT_WEB_APP;
-  payload: ISelected;
+  payload: string;
 }
